% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function for displaying a matrix layout of several plots on one page.}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{}

\item{plotlist}{Combination of plots to display.}

\item{file}{R file containing multiple plots. I guess.}

\item{cols}{Number of columns for layout of plots. Defaults to 1.}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
One-page layout of multiple plots.
}
\description{
Multiple plot function for displaying a matrix layout of several plots on one page.
}

